+++ src/components/SideMenu.vue
@@ -5,11 +5,12 @@
       <h3>
         <font-awesome-icon icon="times-circle" class="fa fa-times-circle" @click.stop="toggleSideMenu" />
         <font-awesome-icon icon="plus-circle" class="fa fa-plus-circle" @click.stop="toggleSideMenu" />
-        <span v-if="!session.isSpectator">{{ t('townsquare.storytellerTools') }}</span>
-        <span v-if="session.isSpectator && grimoire.isNight">{{ t('modal.nightOrder.title') }}</span>
+        <span v-if="!session.sessionId">{{ t('townsquare.disconnected') }}</span>
+        <span v-else-if="!session.isSpectator">{{ t('townsquare.storytellerTools') }}</span>
+        <span v-else-if="session.isSpectator && grimoire.isNight">{{ t('modal.nightOrder.title') }}</span>
       </h3>
 
-      <div v-if="!session.isSpectator">
+      <div v-if="session.sessionId && !session.isSpectator">
         <div class="button-group">
           <button class="button" :disabled="session.gamePhase === 'firstNight'"
             :title="t('menu.grimoire.firstNightSwitch')" @click="setGamePhase('firstNight')">
@@ -24,7 +25,7 @@
             ☾
           </button>
         </div>
-        <div v-if="!grimoire.isNight">
+        <div v-if="!session.isSpectator && !grimoire.isNight">
           <div class="button-group">
             <button @click="setTimer()">
               🕑 {{ timerDuration }} min
@@ -84,6 +85,14 @@
       <div v-if="grimoire.isNight" class="night-order-container">
         <NightOrderTable :night-type="session.gamePhase" />
       </div>
+      <div v-if="!session.sessionId">
+        <div class="button-group">
+          <button @click="hostSession">{{ t('menu.session.storyteller') }}</button>
+        </div>
+        <div class="button-group">
+          <button @click="joinSession">{{ t('menu.session.player') }}</button>
+        </div>
+      </div>
     </div>
   </aside>
 </template>
@@ -120,6 +129,48 @@ const currentTimerType = ref<keyof typeof grimoire.value.timerDurations | null>(
 // Methods
 const toggleSideMenu = () => {
   isSideMenuOpen.value = !isSideMenuOpen.value;
+};
+
+const hostSession = () => {
+  if (session.value.sessionId) return;
+  const sessionId = prompt(
+    t('prompt.createSession'),
+    String(Math.round(Math.random() * 10000)),
+  );
+  if (sessionId) {
+    store.commit('session/clearVoteHistory');
+    store.commit('session/setSpectator', false);
+    store.commit('session/setSessionId', sessionId);
+    store.commit('toggleGrimoire', false);
+    copySessionUrl();
+  }
+};
+
+const joinSession = () => {
+  if (session.value.sessionId) return leaveSession();
+  let sessionId = prompt(t('prompt.joinSession'));
+  if (sessionId && sessionId.match(/^https?:\/\//i)) {
+    sessionId = sessionId.split('#').pop() || null;
+  }
+  if (sessionId) {
+    store.commit('session/clearVoteHistory');
+    store.commit('session/setSpectator', true);
+    store.commit('toggleGrimoire', false);
+    store.commit('session/setSessionId', sessionId);
+  }
+};
+
+const leaveSession = () => {
+  if (confirm(t('prompt.leaveSession'))) {
+    store.commit('session/setSpectator', false);
+    store.commit('session/setSessionId', '');
+  }
+};
+
+const copySessionUrl = () => {
+  const url = window.location.href.split('#')[0];
+  const link = url + '#' + session.value.sessionId;
+  navigator.clipboard.writeText(link);
 };
 
 const setGamePhase = (gamePhase: GamePhase) => {